backend/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
defusedxml==0.7.1
distro==1.9.0
flask==3.1.2
flask-cors==4.0.1
flask-migrate==4.1.0
flask-sqlalchemy==3.1.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.11.0
mako==1.3.10
markupsafe==3.0.2
openai==1.108.0
pip==24.0
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic-core==2.33.2
pymupdf==1.26.4
python-dotenv==1.1.1
requests==2.32.5
sniffio==1.3.1
sqlalchemy==2.0.43
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
werkzeug==3.1.3
youtube-transcript-api==1.2.2
python-jose[cryptography]==3.3.0
